refactoring routes file, user's register, login file

+++ sites/code/views/articles/my_articles.php
@@ -3,6 +3,11 @@
     <h3 style="color: #294c7a;"><?php echo $_SESSION["message"]; ?></h2>
     <?php unset($_SESSION["message"]); ?>
 <?php endif; ?>
-<?php foreach ($articles as $article): ?> 
-    <p>＞＞<a href="?/articles/show/<?php echo $article["id"]; ?>"><?php echo $article["title"]; ?></a>| <a href="?/articles/delete/<?php echo $article["id"]; ?>" style="color: #294c7a; text-decoration: none;">削除</a></p><br>
-<?php endforeach; ?>
+<?php if (isset($error)): ?>
+<p><?php echo htmlspecialchars($error, ENT_QUOTES); ?></p>
+<?php endif; ?>
+<?php if (isset($articles)) :?>
+    <?php foreach ($articles as $article): ?> 
+        <p>＞＞<a href="?/articles/show/<?php echo $article["id"]; ?>"><?php echo $article["title"]; ?></a>| <a href="?/articles/delete/<?php echo $article["id"]; ?>" style="color: #294c7a; text-decoration: none;">削除</a></p><br>
+    <?php endforeach; ?>
+<?php endif; ?>
